+++ assets/GridViewAsset.php
@@ -17,19 +17,25 @@ class GridViewAsset extends AssetBundle
      * @inheritdoc
      */
     public $sourcePath = '@bupy7/grid/resources';
-    public $js = [
-        'js/resizable-columns.min.js',
-    ];
+    public $js = [];
     /**
      * @inheritdoc
      */
-    public $css = [
-        'css/resizable-columns.min.css',
-    ];
+    public $css = [];
     /**
      * @inheritdoc
      */
     public $depends = [
         'yii\grid\GridViewAsset',
     ];
+    
+    /**
+     * @inheritdoc
+     */
+    public function init()
+    {
+        parent::init();
+        $this->css[] = 'css/resizable-columns' . (!YII_DEBUG ? '.min' : '') . '.css';
+        $this->js[] = 'js/resizable-columns' . (!YII_DEBUG ? '.min' : '') . '.js';
+    }
 }